Wavefronts.jl documentation

Example of using Wavefronts.jl

Making interference patterns

Retrieving the phase step by step

Automated method

Wavefronts.jl documentation

Functions

Types

Index

Functions

Wavefronts.retrievephase — Function

retrievephase(intensity0, intensity1, intensity2, intensity3; correct_aberrations, mask)

Basically compose rawphase and unwrapphase. If correct_aberrations is given, tries to correct the given aberrations with the given mask.

See also: rawphase, unwrapphase, correct

Wavefronts.rawphase — Function

rawphase(intensity0, intensity1, intensity2, intensity3)

Compute the raw phase from four intensity of interference pattern. The reference in the pattern of intensity intensityi is assumed to be shifted of π/2.

See also: unwrapphase, retrievephase

Wavefronts.unwrapphase — Function

unwrapphase(wrappedphase)

Unwrap the phase. The method used is inspired by the Standard Experiment Laboratory script from the MSc in Optics and Photonics at Imperial College London. The input is typically the output of rawphase.

See also: rawphase, retrievephase

Wavefronts.correct — Function

correct(aberration, data; mask=(x,y)->1)

Correct the given aberration in the data by projecting the data on the aberration.

See also: project

Wavefronts.project — Function

project(aberration, data; mask=(x,y)->1)

Project the data onto the Zernike polynomial corresponding to the aberration in order to get the fitting Aberration. A mask can be used to fit only in certain parts of the data (typically usefull if your interference pattern is only relevant in a circular region). The coordinates are assumed to be normalized such that (x,y) ∈ [-1;1]².

See also: correct

Wavefronts.intensity — Function

intensity(aberration, coordinates=:cartesian; mask=(x,y)->1)

Creates a function mapping the intensity in the given coordinates system (:cartesian or :polar). This is equivalent of building two Wavefront, one with no aberration and the other with the given aberration.

See also: phase, amplitude

intensity(wavefront)

Create a function mapping the intensity in the coordinates of the wavefront.

See also: field, amplitude, phase

intensity(wavefront, reference)

Create a function mapping the intensity of the interference pattern of the wavefront with the reference. Both should be in the same coordinate system.

See also: field, amplitude, phase

Wavefronts.field — Function

field(wavefront)

Create a function mapping the field in the coordinates of the wavefront.

See also: intensity, amplitude, phase

Wavefronts.phase — Function

phase(aberration, coordinates=:cartesian)

Create a function mapping the phase of the aberration in the given coordinates system (:cartesian or :polar).

See also: intensity, amplitude

phase(wavefront)

Create a function mapping the phase of the wavefront.

See also: intensity, amplitude, field

Wavefronts.amplitude — Function

amplitude(wavefront)

Create a function mapping the amplitude of the wavefront.

See also: intensity, field, phase

Wavefronts.circ — Function

circ(x,y)

Circular mask of radius 1.

Types

Aberrations

Wavefronts.Aberration — Type

Represents an aberration. This uses Zernike polynomials.

See also: Piston, Tilt, Tip, Defocus, ObliqueAstigmatism, VerticalAstigmatism, HorizontalComa, VerticalComa, ObliqueTrefoil, VerticalTrefoil, Spherical, AddAberration, Astigmatism, Coma, Trefoil

Wavefronts.Piston — Type

Piston aberration.

Wavefronts.Tilt — Type

Tilt aberration.

Wavefronts.Tip — Type

Tip aberration.

Wavefronts.Defocus — Type

Defocus aberration.

Wavefronts.ObliqueAstigmatism — Type

ObliqueAstigmatism aberration.

Wavefronts.VerticalAstigmatism — Type

VerticalAstigmatism aberration.

Wavefronts.HorizontalComa — Type

HorizontalComa aberration.

Wavefronts.VerticalComa — Type

VerticalComa aberration.

Wavefronts.ObliqueTrefoil — Type

ObliqueTrefoil aberration.

Wavefronts.VerticalTrefoil — Type

VerticalTrefoil aberration.

Wavefronts.Spherical — Type

Spherical aberration.

Wavefronts.AddAberration — Type

The addition of two aberrations. It is created when + operator is used on two Aberration.

Wavefronts.Astigmatism — Type

Astigmatism aberration.

Wavefronts.Coma — Type

Coma aberration.

Wavefronts.Trefoil — Type

Trefoil aberration.

Wavefronts

Wavefronts.Wavefront — Type

Represents a wavefront.

Wavefronts.AddWavefront — Type

Represents the addition of a two wavefronts. It is created when + operator is used on two Wavefront.

Index

Wavefronts.Aberration

Wavefronts.AddAberration

Wavefronts.AddWavefront

Wavefronts.Astigmatism

Wavefronts.Coma

Wavefronts.Defocus

Wavefronts.HorizontalComa

Wavefronts.ObliqueAstigmatism

Wavefronts.ObliqueTrefoil

Wavefronts.Piston

Wavefronts.Spherical

Wavefronts.Tilt

Wavefronts.Tip

Wavefronts.Trefoil

Wavefronts.VerticalAstigmatism

Wavefronts.VerticalComa

Wavefronts.VerticalTrefoil

Wavefronts.Wavefront

Wavefronts.amplitude

Wavefronts.circ

Wavefronts.correct

Wavefronts.field

Wavefronts.intensity

Wavefronts.phase

Wavefronts.project

Wavefronts.rawphase

Wavefronts.retrievephase

Wavefronts.unwrapphase